// frontend/models/user.ts
export class User{
  username: String = ""
  password: String = ""
  type: String = ""
  name: String = ""
  lastname: String = ""
  gender: String = ""
  address: String = ""
  phone_number: String = ""
  mail: String = ""
  credit_card_number: String = ""
  picture: String | ArrayBuffer | null = null
  blokiran: Number = 0
}